+++ android/app/src/main/java/com/badgermobile/MainActivity.java
@@ -1,4 +1,4 @@
-package com.badgermobile;
+package com.optnwallet;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "badgerMobile";
+    return "optnwallet";
   }
 }