+++ android/app/src/main/java/com/badgermobile/MainApplication.java
@@ -1,4 +1,4 @@
-package com.badgermobile;
+package com.optnwallet;
 
 import android.app.Application;
 import android.content.Context;
@@ -49,7 +49,7 @@ public class MainApplication extends Application implements ReactApplication {
     SoLoader.init(this, /* native exopackage */ false);
     initializeFlipper(this); // Remove this line if you don't want Flipper enabled
   }
-  
+
   /**
    * Loads Flipper in React Native templates.
    *