+++ components/ContractFunctionsFactory/index.tsx
@@ -1,7 +1,9 @@
 import P2pkhFunctionsFactory from "./P2pkh";
+/*
 import SLPGenesisFunctionsFactory from "./SlpGenesis";
 import SLPMintFunctionsFactory from "./SlpMint";
 import SLPSendFunctionsFactory from "./SlpSend";
+*/
 import Bip38FunctionsFactory from "./Bip38";
 
 const ContractFunctionsFactory = (contractName: string, fnName: string) => {
@@ -11,6 +13,7 @@ const ContractFunctionsFactory = (contractName: string, fnName: string) => {
     case "P2PKH":
       functionsFactory = P2pkhFunctionsFactory;
       break;
+    /*
     case "SLPGenesis":
       functionsFactory = SLPGenesisFunctionsFactory;
       break;
@@ -20,11 +23,12 @@ const ContractFunctionsFactory = (contractName: string, fnName: string) => {
     case "SLPSend":
       functionsFactory = SLPSendFunctionsFactory;
       break;
+*/
     case "Bip38":
       functionsFactory = Bip38FunctionsFactory;
       break;
     default:
-      return null
+      return null;
   }
 
   const built = functionsFactory.build(fnName);