+++ data/restURLs/constants.ts
@@ -1,0 +1,5 @@
+export const ADD_NODEINFO = "badger::restURL:ADD_NODEINFO";
+
+export const GET_NODEINFO_START = "badger::restURL:GET_NODEINFO_START";
+export const GET_NODEINFO_SUCCESS = "badger::restURL:GET_NODEINFO_SUCCESS";
+export const GET_NODEINFO_FAIL = "badger::restURL:GET_NODEINFO_FAIL";