+++ screens/HomeScreen.tsx
@@ -9,7 +9,8 @@ import {
   StyleSheet,
   SectionList,
   View,
-  TouchableOpacity
+  TouchableOpacity,
+  Image
 } from "react-native";
 import { NavigationScreenProps } from "react-navigation";
 import uuidv5 from "uuid/v5";
@@ -40,6 +41,8 @@ import {
   formatFiatAmount,
   computeFiatAmount
 } from "../utils/balance-utils";
+
+import OPTNWelcome1 from "../assets/images/OPTNWelcome1.png";
 
 const SECOND = 1000;
 
@@ -151,7 +154,9 @@ const HomeScreen = ({
     updateUtxos(address, addressSlp);
     const utxoInterval = setInterval(
       () => updateUtxos(address, addressSlp),
-      15 * SECOND
+      // Fullstack.cash rate limit set defaults at 3 requests per minute
+      // 15 * SECOND
+      60 * SECOND
     );
     return () => {
       clearInterval(utxoInterval);
@@ -168,7 +173,7 @@ const HomeScreen = ({
     // Update transaction history interval
     const transactionInterval = setInterval(() => {
       updateTransactions(address, addressSlp);
-    }, 30 * SECOND);
+    }, 60 * SECOND);
     return () => {
       clearInterval(transactionInterval);
     };
@@ -287,7 +292,8 @@ const HomeScreen = ({
     <SafeAreaView>
       <View
         style={{
-          height: "100%"
+          height: "100%",
+          alignItems: "stretch"
         }}
       >
         <ScrollView
@@ -312,6 +318,16 @@ const HomeScreen = ({
           ) : (
             <Spacer large />
           )}
+          <Image
+            source={OPTNWelcome1}
+            style={{
+              width: 350,
+              height: 150,
+              resizeMode: "contain",
+              alignItems: "center"
+            }}
+          />
+          {/*
           <H1 center spacing="loose" weight="bold">
             Badger
           </H1>
@@ -320,11 +336,18 @@ const HomeScreen = ({
             BCH and SLP wallet
           </T>
           <Spacer />
+*/}
           <View
             style={{
               position: "relative"
             }}
           >
+            <Button
+              onPress={() => {
+                navigation.navigate("ContractListScreen");
+              }}
+              text="Cashscript  Contracts"
+            />
             <SectionList
               sections={walletSections}
               renderSectionHeader={({ section }) => (
@@ -351,12 +374,6 @@ const HomeScreen = ({
                 )
               }
               keyExtractor={(item, index) => `${index}`}
-            />
-            <Button
-              onPress={() => {
-                navigation.navigate("ContractListScreen");
-              }}
-              text="Contracts"
             />
             <Spacer />
             {!initialLoadingDone && (