+++ utils/balance-utils.ts
@@ -125,17 +125,15 @@ const getHistoricalSlpTransactions = async (
               "in.e.a": address.slice(12)
             },
             {
-              "slp.detail.outputs.address": bchjs.SLP.Address.toSLPAddress(
-                address
-              )
+              //            "slp.detail.outputs.address": bchjs.SLP.Address.toSLPAddress(
+              "slp.detail.outputs.address": SLP.Address.toSLPAddress(address)
             },
             {
               "in.e.a": SLP.Address.toSLPAddress(addressSlp)
             },
             {
-              "slp.detail.outputs.address": bchjs.SLP.Address.toSLPAddress(
-                addressSlp
-              )
+              //            "slp.detail.outputs.address": bchjs.SLP.Address.toSLPAddress(
+              "slp.detail.outputs.address": SLP.Address.toSLPAddress(addressSlp)
             }
           ],
           "slp.valid": true,