+++ utils/schemeParser-utils.ts
@@ -1,7 +1,6 @@
 import BigNumber from "bignumber.js";
 import { Utils } from "slpjs";
 import { SLP } from "./slp-sdk-utils";
-import { bchjs } from "./bch-js-utils";
 
 import { TokenData } from "../data/tokens/reducer";
 
@@ -15,7 +14,7 @@ const parseAddress = (address: string) => {
       throw new Error("invalid address");
     }
 
-    return bchjs.SLP.Address.toCashAddress(address);
+    return SLP.Address.toCashAddress(address);
   } else if (type === "slpaddr") {
     try {
       checkIsValid("slpaddr", address);
@@ -23,7 +22,7 @@ const parseAddress = (address: string) => {
       throw new Error("invalid address");
     }
 
-    return bchjs.SLP.Address.toSLPAddress(address);
+    return SLP.Address.toSLPAddress(address);
   }
   return address;
 };
@@ -104,7 +103,7 @@ const parseBCHScheme = (
   } catch (error) {
     throw new Error("invalid address");
   }
-  const cashAddress = bchjs.SLP.Address.toCashAddress(addressFromScheme);
+  const cashAddress = SLP.Address.toCashAddress(addressFromScheme);
 
   const amount = getValue(scheme, "amount");
   const parsedAmount = parseAmount(amount);
@@ -184,14 +183,14 @@ const parseAmount = (value?: string | number | null): string | null => {
 };
 
 const getType = (address: string) => {
-  return bchjs.SLP.Address.detectAddressFormat(address);
+  return SLP.Address.detectAddressFormat(address);
 };
 
 const checkIsValid = (type: string, address: string) => {
   if (type === "cashaddr") {
-    return bchjs.SLP.Address.isCashAddress(address);
+    return SLP.Address.isCashAddress(address);
   } else if (type === "slpaddr") {
-    return bchjs.SLP.Address.isSLPAddress(address);
+    return SLP.Address.isSLPAddress(address);
   } else {
     console.warn("error in checkisvalid");
   }