+++ utils/transaction-utils.ts
@@ -5,7 +5,6 @@ import { ECPair } from "../data/accounts/reducer";
 import { TokenData } from "../data/tokens/reducer";
 
 import { SLP } from "./slp-sdk-utils";
-import { bchjs } from "./bch-js-utils";
 
 const slpjs = require("slpjs");
 
@@ -47,14 +46,13 @@ const getSLPTxType = (scriptASMArray: string[]) => {
 };
 
 const getAllUtxo = async (address: string) => {
-  const result = await bchjs.Electrumx.utxo(address);
+  const result = await SLP.Address.utxo(address);
   return result.utxos;
 };
 
 const getTransactionDetails = async (txid: string | string[]) => {
-  console.log(txid);
   try {
-    const result = await bchjs.SLP.Utils.txDetails(txid);
+    const result = await SLP.Transaction.details(txid);
     return result;
   } catch (e) {
     throw e;
@@ -71,7 +69,7 @@ const decodeTxOut = (txOut: UTXO) => {
 
   const vout = parseInt(txOut.vout, 10);
 
-  const script = bchjs.Script.toASM(
+  const script = SLP.Script.toASM(
     Buffer.from(txOut.tx.vout[0].scriptPubKey.hex, "hex")
   ).split(" ");
 
@@ -146,7 +144,7 @@ const decodeTxOut = (txOut: UTXO) => {
 
 // Straight from Badger plugin
 const decodeTokenMetadata = (txDetails: UTXO): TokenData => {
-  const script = bchjs.Script.toASM(
+  const script = SLP.Script.toASM(
     Buffer.from(txDetails.vout[0].scriptPubKey.hex, "hex")
   ).split(" ");
   const type = getSLPTxType(script);
@@ -167,7 +165,7 @@ const decodeTokenMetadata = (txDetails: UTXO): TokenData => {
 };
 
 const encodeOpReturn = async (dataArray: string[]) => {
-  const script = [bchjs.Script.opcodes.OP_RETURN];
+  const script = [SLP.Script.opcodes.OP_RETURN];
 
   dataArray.forEach(data => {
     if (typeof data === "string" && data.substring(0, 2) === "0x") {
@@ -177,11 +175,11 @@ const encodeOpReturn = async (dataArray: string[]) => {
     }
   });
 
-  return await bchjs.Script.encode(script);
+  return await SLP.Script.encode(script);
 };
 
 const publishTx = async (hex: string) => {
-  const result = await bchjs.RawTransactions.sendRawTransaction(hex);
+  const result = await SLP.RawTransactions.sendRawTransaction(hex);
 
   try {
     if (result.length === 64) {
@@ -210,7 +208,7 @@ const signAndPublishBchTransaction = async (
       ? await encodeOpReturn(opReturn.data)
       : null;
 
-    const transactionBuilder = new bchjs.TransactionBuilder("mainnet");
+    const transactionBuilder = new SLP.TransactionBuilder("mainnet");
     const inputUtxos = [];
     let byteCount = 0;
     let totalUtxoAmount = 0;
@@ -224,7 +222,7 @@ const signAndPublishBchTransaction = async (
       totalUtxoAmount += utxo.satoshis;
       inputUtxos.push(utxo);
 
-      byteCount = bchjs.BitcoinCash.getByteCount(
+      byteCount = SLP.BitcoinCash.getByteCount(
         {
           P2PKH: inputUtxos.length
         },
@@ -374,7 +372,7 @@ const signAndPublishSlpTransaction = async (
     }
   }
 
-  const transactionBuilder = new bchjs.TransactionBuilder("mainnet");
+  const transactionBuilder = new SLP.TransactionBuilder("mainnet");
   let totalUtxoAmount = 0;
 
   inputUtxos.forEach(utxo => {
@@ -478,21 +476,21 @@ const getPaperKeypair = async (wif?: string | null) => {
     );
   }
 
-  const keypair = await bchjs.ECPair.fromWIF(wif);
+  const keypair = await SLP.ECPair.fromWIF(wif);
   return keypair;
 };
 
 const getPaperUtxos = async (keypair: ECPair): Promise<UtxosByKey> => {
   try {
     // Generate the public address associated with the private key.
-    const fromAddr: string = bchjs.ECPair.toCashAddress(keypair);
+    const fromAddr: string = SLP.ECPair.toCashAddress(keypair);
 
     // Get UTXOs associated with public address.
-    const u = await bchjs.Electrumx.utxo(fromAddr);
+    const u = await SLP.Address.utxo(fromAddr);
     const utxosAll = u.utxos as PaperUTXO[];
 
     let utxosDetails = [] as PaperUTXO[];
-    utxosDetails = await bchjs.SLP.Util.tokenUtxoDetails(utxosAll);
+    utxosDetails = await SLP.Util.tokenUtxoDetails(utxosAll);
     console.assert(
       utxosAll.length === utxosDetails.length,
       "UTXO Details and UTXOs differ in length"
@@ -556,10 +554,10 @@ const sweepPaperWallet = async (
     const hasBCH = paperBalanceKeys.includes("BCH");
 
     // Generate a keypair from the WIF
-    const keyPair = bchjs.ECPair.fromWIF(wif);
-    const fromAddr: string = bchjs.ECPair.toCashAddress(keyPair);
-    const transactionBuilder = new bchjs.TransactionBuilder(
-      bchjs.Address.detectAddressNetwork(fromAddr)
+    const keyPair = SLP.ECPair.fromWIF(wif);
+    const fromAddr: string = SLP.ECPair.toCashAddress(keyPair);
+    const transactionBuilder = new SLP.TransactionBuilder(
+      SLP.Address.detectAddressNetwork(fromAddr)
     );
 
     if (tokenId && tokenDecimals != null && hasBCH) {
@@ -633,7 +631,7 @@ const sweepPaperWallet = async (
       }
 
       // get byte count to calculate fee. paying 1.1 sat/byte
-      const byteCount: number = bchjs.BitcoinCash.getByteCount(
+      const byteCount: number = SLP.BitcoinCash.getByteCount(
         { P2PKH: bchUtxos.length },
         { P2PKH: 1 }
       );
@@ -644,7 +642,7 @@ const sweepPaperWallet = async (
 
       // add output w/ address and amount to send
       transactionBuilder.addOutput(
-        bchjs.Address.toLegacyAddress(addressBch),
+        SLP.Address.toLegacyAddress(addressBch),
         sendAmount
       );
 
@@ -667,7 +665,7 @@ const sweepPaperWallet = async (
       const hex: string = tx.toHex();
 
       // Broadcast the transaction to the BCH network.
-      txid = await bchjs.RawTransactions.sendRawTransaction(hex);
+      txid = await SLP.RawTransactions.sendRawTransaction(hex);
       return txid;
     } else if (tokenId && tokenDecimals != null && !hasBCH) {
       // SLP only sweep